Renamed test
