+++ tests/Feature/Maintenances/Api/EditMaintenanceTest.php
@@ -1,69 +1,64 @@
 <?php
 
-namespace Tests\Feature\AssetMaintenances\Ui;
+namespace Tests\Feature\Maintenances\Api;
 
 use App\Models\Asset;
-use App\Models\AssetMaintenance;
+use App\Models\Maintenance;
 use App\Models\Supplier;
 use App\Models\User;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 use Tests\TestCase;
 
-class CreateAssetMaintenanceTest extends TestCase
+class EditMaintenanceTest extends TestCase
 {
     public function testPageRenders()
     {
         $this->actingAs(User::factory()->superuser()->create())
-            ->get(route('maintenances.create'))
+            ->get(route('maintenances.update', Maintenance::factory()->create()->id))
             ->assertOk();
     }
 
 
-    public function testCanCreateAssetMaintenance()
+    public function testCanEditMaintenance()
     {
         Storage::fake('public');
         $actor = User::factory()->superuser()->create();
-
         $asset = Asset::factory()->create();
         $supplier = Supplier::factory()->create();
+        $maintenance = Maintenance::factory()->create();
 
-        $this->actingAs($actor)
+        $response = $this->actingAs($actor)
             ->followingRedirects()
-            ->post(route('maintenances.store'), [
-                'title' => 'Test Maintenance',
-                'selected_assets' => [$asset->id],
+            ->patch(route('maintenances.update',  $maintenance), [
+                'name' => 'Test Maintenance',
                 'supplier_id' => $supplier->id,
                 'asset_maintenance_type' => 'Maintenance',
                 'start_date' => '2021-01-01',
                 'completion_date' => '2021-01-10',
                 'is_warranty' => '1',
-                'cost' => '100.00',
                 'image' => UploadedFile::fake()->image('test_image.png'),
                 'notes' => 'A note',
             ])
             ->assertOk();
 
-        // Since we rename the file in the ImageUploadRequest, we have to fetch the record from the database
-        $assetMaintenance = AssetMaintenance::where('title', 'Test Maintenance')->first();
+        $this->followRedirects($response)->assertSee('alert-success');
 
+        $maintenance->refresh();
         // Assert file was stored...
-        Storage::disk('public')->assertExists(app('asset_maintenances_path').$assetMaintenance->image);
+        Storage::disk('public')->assertExists(app('maintenances_path').$maintenance->image);
 
 
         $this->assertDatabaseHas('asset_maintenances', [
-            'asset_id' => $asset->id,
             'supplier_id' => $supplier->id,
             'asset_maintenance_type' => 'Maintenance',
-            'title' => 'Test Maintenance',
+            'name' => 'Test Maintenance',
             'is_warranty' => 1,
             'start_date' => '2021-01-01',
             'completion_date' => '2021-01-10',
             'asset_maintenance_time' => '9',
             'notes' => 'A note',
-            'cost' => '100.00',
-            'image' => $assetMaintenance->image,
-            'created_by' => $actor->id,
+            'image' => $maintenance->image,
         ]);
     }
 }
