+++ tests/Feature/Maintenances/Ui/CreateMaintenanceTest.php
@@ -1,40 +1,38 @@
 <?php
 
-namespace Tests\Feature\AssetMaintenances\Api;
+namespace Tests\Feature\Maintenances\Ui;
 
 use App\Models\Asset;
-use App\Models\AssetMaintenance;
-use App\Models\Category;
+use App\Models\Maintenance;
 use App\Models\Supplier;
 use App\Models\User;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
-use Illuminate\Testing\Fluent\AssertableJson;
 use Tests\TestCase;
 
-class CreateAssetMaintenanceTest extends TestCase
+class CreateMaintenanceTest extends TestCase
 {
-
-
-    public function testRequiresPermissionToCreateAssetMaintenance()
+    public function testPageRenders()
     {
-        $this->actingAsForApi(User::factory()->create())
-            ->postJson(route('api.maintenances.store'))
-            ->assertForbidden();
+        $this->actingAs(User::factory()->superuser()->create())
+            ->get(route('maintenances.create'))
+            ->assertOk();
     }
-    public function testCanCreateAssetMaintenance()
-    {
 
+
+    public function testCanCreateMaintenance()
+    {
         Storage::fake('public');
         $actor = User::factory()->superuser()->create();
 
         $asset = Asset::factory()->create();
         $supplier = Supplier::factory()->create();
 
-        $response = $this->actingAsForApi($actor)
-            ->postJson(route('api.maintenances.store'), [
-                'title' => 'Test Maintenance',
-                'asset_id' => $asset->id,
+        $this->actingAs($actor)
+            ->followingRedirects()
+            ->post(route('maintenances.store'), [
+                'name' => 'Test Maintenance',
+                'selected_assets' => [$asset->id],
                 'supplier_id' => $supplier->id,
                 'asset_maintenance_type' => 'Maintenance',
                 'start_date' => '2021-01-01',
@@ -44,29 +42,28 @@ class CreateAssetMaintenanceTest extends TestCase
                 'image' => UploadedFile::fake()->image('test_image.png'),
                 'notes' => 'A note',
             ])
-            ->assertOk()
-            ->assertStatus(200);
+            ->assertOk();
 
         // Since we rename the file in the ImageUploadRequest, we have to fetch the record from the database
-        $assetMaintenance = AssetMaintenance::where('title', 'Test Maintenance')->first();
+        $assetMaintenance = Maintenance::where('title', 'Test Maintenance')->first();
 
         // Assert file was stored...
-        Storage::disk('public')->assertExists(app('asset_maintenances_path').$assetMaintenance->image);
+        Storage::disk('public')->assertExists(app('maintenances_path').$assetMaintenance->image);
+
 
         $this->assertDatabaseHas('asset_maintenances', [
             'asset_id' => $asset->id,
             'supplier_id' => $supplier->id,
             'asset_maintenance_type' => 'Maintenance',
-            'title' => 'Test Maintenance',
+            'name' => 'Test Maintenance',
             'is_warranty' => 1,
             'start_date' => '2021-01-01',
             'completion_date' => '2021-01-10',
+            'asset_maintenance_time' => '9',
             'notes' => 'A note',
+            'cost' => '100.00',
             'image' => $assetMaintenance->image,
             'created_by' => $actor->id,
         ]);
     }
-
-
-
 }
