+++ tests/Feature/Maintenances/Ui/EditMaintenanceTest.php
@@ -1,64 +1,65 @@
 <?php
 
-namespace Tests\Feature\AssetMaintenances\Api;
+namespace Tests\Feature\Maintenances\Ui;
 
 use App\Models\Asset;
-use App\Models\AssetMaintenance;
+use App\Models\Maintenance;
 use App\Models\Supplier;
 use App\Models\User;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 use Tests\TestCase;
 
-class EditAssetMaintenanceTest extends TestCase
+class EditMaintenanceTest extends TestCase
 {
     public function testPageRenders()
     {
         $this->actingAs(User::factory()->superuser()->create())
-            ->get(route('maintenances.update', AssetMaintenance::factory()->create()->id))
+            ->get(route('maintenances.edit', Maintenance::factory()->create()->id))
             ->assertOk();
     }
 
-
-    public function testCanEditAssetMaintenance()
+    public function testCanUpdateMaintenance()
     {
-        Storage::fake('public');
         $actor = User::factory()->superuser()->create();
         $asset = Asset::factory()->create();
+        $assetMaintenance = Maintenance::factory()->create(['asset_id' => $asset]);
         $supplier = Supplier::factory()->create();
-        $maintenance = AssetMaintenance::factory()->create();
 
-        $response = $this->actingAs($actor)
+        $this->actingAs($actor)
             ->followingRedirects()
-            ->patch(route('maintenances.update',  $maintenance), [
-                'title' => 'Test Maintenance',
+            ->put(route('maintenances.update', $assetMaintenance), [
+                'name' => 'Test Maintenance',
+                'asset_id' => $asset->id,
                 'supplier_id' => $supplier->id,
                 'asset_maintenance_type' => 'Maintenance',
                 'start_date' => '2021-01-01',
                 'completion_date' => '2021-01-10',
-                'is_warranty' => '1',
+                'is_warranty' => 1,
                 'image' => UploadedFile::fake()->image('test_image.png'),
+                'cost' => '100.99',
                 'notes' => 'A note',
             ])
             ->assertOk();
 
-        $this->followRedirects($response)->assertSee('alert-success');
+        // Since we rename the file in the ImageUploadRequest, we have to fetch the record from the database
+        $assetMaintenance = Maintenance::where('title', 'Test Maintenance')->first();
 
-        $maintenance->refresh();
         // Assert file was stored...
-        Storage::disk('public')->assertExists(app('asset_maintenances_path').$maintenance->image);
-
+        Storage::disk('public')->assertExists(app('maintenances_path').$assetMaintenance->image);
 
         $this->assertDatabaseHas('asset_maintenances', [
+            'asset_id' => $asset->id,
             'supplier_id' => $supplier->id,
             'asset_maintenance_type' => 'Maintenance',
-            'title' => 'Test Maintenance',
+            'name' => 'Test Maintenance',
             'is_warranty' => 1,
             'start_date' => '2021-01-01',
             'completion_date' => '2021-01-10',
             'asset_maintenance_time' => '9',
             'notes' => 'A note',
-            'image' => $maintenance->image,
+            'cost' => '100.99',
         ]);
     }
+
 }
